Fixed long line and seperated methods into seperate blocks
